backend/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
numpy>=1.21.0
#opencv-python>=4.5.0

# Data handling
pandas>=1.3.0
pyyaml>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.64.0

# Data augmentation
albumentations>=1.3.0

# Optional (for visualization)
# tensorboard>=2.8.0
# wandb>=0.13.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.3.0
isort>=5.10.1
flake8>=4.0.0
mypy>=0.950.0